+++ src/client/grid/canvasGrid/Grid.tsx
@@ -8,6 +8,8 @@ import { ColorInfo, clearTile, renderTile } from './tile';
 import GridData from '@logic-pad/core/data/grid';
 import TileConnections from '@logic-pad/core/data/tileConnections';
 
+const MAX_SIZE = 5000;
+
 interface GridRenderData {
   grid: GridData;
   connections: TileConnections[][];
@@ -62,9 +64,20 @@ export default memo(function Grid({
     [theme]
   );
 
+  const scale = useMemo(() => {
+    if (grid.width * size <= MAX_SIZE && grid.height * size <= MAX_SIZE) {
+      return 1;
+    }
+    return Math.min(
+      MAX_SIZE / (grid.width * size),
+      MAX_SIZE / (grid.height * size)
+    );
+  }, [grid.width, grid.height, size]);
+
   useEffect(() => {
     canvasCtx.current ??= canvasRef.current?.getContext('2d') ?? null;
     if (!canvasCtx.current) return;
+    const renderSize = size * scale;
     const ctx = canvasCtx.current;
     for (let y = 0; y < grid.height; y++) {
       for (let x = 0; x < grid.width; x++) {
@@ -72,32 +85,47 @@ export default memo(function Grid({
         const oldTile = prevData.current?.grid.getTile(x, y);
         if (
           prevData.current?.colorInfo === colorInfo &&
-          prevData.current?.size === size &&
+          prevData.current?.size === renderSize &&
           prevData.current?.grid.width === grid.width &&
           prevData.current?.grid.height === grid.height &&
           oldTile?.equals(tile) &&
           prevData.current?.connections[y]?.[x]?.equals(tileConnections[y][x])
         )
           continue;
-        clearTile(ctx, x, y, size);
+        clearTile(ctx, x, y, renderSize);
         if (!tile.exists) continue;
-        renderTile(ctx, x, y, size, tile, tileConnections[y][x], colorInfo);
+        renderTile(
+          ctx,
+          x,
+          y,
+          renderSize,
+          tile,
+          tileConnections[y][x],
+          colorInfo
+        );
       }
     }
     prevData.current = {
       grid,
       connections: tileConnections,
-      size,
+      size: renderSize,
       colorInfo,
     };
-  }, [grid, size, canvasCtx, colorInfo, tileConnections]);
+  }, [grid, size, canvasCtx, colorInfo, tileConnections, scale]);
 
   return (
     <div className={cn('relative', className)} style={containerStyle}>
       <canvas
         ref={canvasRef}
-        width={grid.width * size}
-        height={grid.height * size}
+        width={grid.width * size * scale}
+        height={grid.height * size * scale}
+        style={useMemo(
+          () => ({
+            transformOrigin: 'top left',
+            transform: `scale(${1 / scale})`,
+          }),
+          [scale]
+        )}
         className="absolute inset-0"
       />
       <PointerCaptureOverlay